Integrate Spawner, make example scene
Integrated spawner with editor tools, spawning prefabs, etc
Made example scene demonstrating SpawnManager and Spawner
