+++ Assets/Developers/Hazel/EvenLitter.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd5634cf6af5dc44c9c95f2c5dd176f4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
