+++ Assets/Developers/Hazel/TestSpawnedThing Variant.prefab
@@ -1,0 +1,75 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &2367190151400736876
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 4548827022190203505, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_Name
+      value: TestSpawnedThing Variant
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203517, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4548827022190203519, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: fd5634cf6af5dc44c9c95f2c5dd176f4, type: 2}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 27d1f4605b510cf4b8b7f26e8672ed35, type: 3}
