+++ Assets/Developers/Hazel/TestSpawnedThing Variant.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 05e90e899db3b044b94041c44053a6ef
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
