+++ Assets/Developers/Hazel/TestSpawnedThing.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 27d1f4605b510cf4b8b7f26e8672ed35
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
