+++ Assets/Resources/Data/Spawners.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e076b3e7fe7b9a942a7b5f407e17ec74
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
