+++ Assets/Resources/Data/Spawners/Neutral.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bf135efcdc1e04149b6d6fab5a127132
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
