+++ Assets/Resources/Data/Spawners/Powerup.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59f31261aa077354f98cd8401be50c9d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
