--- Assets/Resources/Data/Spawnpoints.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e076b3e7fe7b9a942a7b5f407e17ec74
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
