--- Assets/Resources/Data/Spawnpoints/Enemy.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2224c49450730c24f80de7b4d26c6cac
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
