--- Assets/Resources/Data/Spawnpoints/Friendly.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1b6e7e1c1902ab44e979d44d5e386e6c, type: 3}
-  m_Name: Friendly
-  m_EditorClassIdentifier: 
-  discThickness: 4
-  discColor: {r: 0.38581347, g: 0.9622642, b: 0.7758562, a: 1}
