--- Assets/Resources/Data/Spawnpoints/Friendly.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a73c2b10e9de1ef458b0b04909223ce6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
