--- Assets/Resources/Data/Spawnpoints/Neutral.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1b6e7e1c1902ab44e979d44d5e386e6c, type: 3}
-  m_Name: Neutral
-  m_EditorClassIdentifier: 
-  discThickness: 2
-  discColor: {r: 0.68773144, g: 0.66202384, b: 0.8207547, a: 1}
-  hoverDiscColor: {r: 1, g: 1, b: 0.4745098, a: 1}
