+++ Assets/Scenes/Examples/SpawnerExample.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3894a3d3bacd54d598ed70b6ef055746
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
