--- Assets/Scripts/Data/SpawnpointEditorPreviewData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b6e7e1c1902ab44e979d44d5e386e6c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
