+++ Assets/Scripts/Editor/SpawnerEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b57b6ef2bf3095a47b16d8a951155d9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
