--- Assets/Scripts/Editor/SpawnpointManagerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e001d12a3804d8e4f87e35da62e6dfa5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
