+++ Assets/Scripts/Gameplay/Spawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 873097eb908f8415fa3437589ef8f3ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
