--- Assets/Scripts/Gameplay/Spawnpoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cf749d4c91c22bc4b919e8ae0b28c77c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
