--- Assets/Scripts/Management/SpawnpointManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: caa67121876c0e34a9710b75a1bcbc00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
